% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSensitivity.R
\name{findSensitivity}
\alias{findSensitivity}
\title{Make predictions about the chemoresistance for PDAC samples based on RNA-seq gene expression levels}
\usage{
findSensitivity(Sample_Expression_Frame)
}
\arguments{
\item{Sample_Expression_Frame}{A dataframe of normalized gene expression values for each PDAC sample. Rows should represent genes with row names containing the gene names. Columns should represent samples with column names providing the sample names. It is important that the gene expression are normalized for read depth.}
}
\description{
Make predictions about the chemoresistance for PDAC samples based on RNA-seq gene expression levels
}
\examples{
data("CCLE_Share")
findSensitivity(CCLE_Share)
}

